tests: make souphttpsrc unit test work even if http_proxy is set
We're testing with an http server on localhost, but don't support
an exception list for the http_proxy, so just unset the environment
variable to make sure we can run this test properly even if the
environment has http_proxy set.

Also, don't skip all tests if there is an issue with the SSL server,
just run the non-SSL tests then.

https://jenkins.qa.ubuntu.com/view/Raring/view/JHBuild%20Gnome/job/jhbuild-amd64-gst-plugins-good/