tests: fix skipped rganalysis tests
In 0.10 elements would post tag messages on the bus
directly, and rganalysis would only post a tag message
when it changed tags. In 1.0, only sinks post tag
messages when they receive the serialised tag event.
This means that we get an additional tag message on
the bus now where we didn't expect one before.

https://bugzilla.gnome.org/show_bug.cgi?id=695090