bluez: Remove a2dpsrc bin
This element isn't really useful. The original plan was to use this to
eventually allow dynamically picking a format based on what we can
support while negotiating with the source device. We can just plug in a
decodebin downstream and it'll handle RTP depayloading and decoding of
whatever format we negotiated with the source device.