rtsp: calculate the local ip address in accept
Calculate the local IP address in the accept call. We need to place this IP
address in the GET reply in the X-Server-IP-Address header so that the client
knows where to send the POST to in case of tunneled RTSP. Before this patch
it used the client IP address, which would make the client send the POST request
to itself and fail.

Fixes https://bugzilla.gnome.org/show_bug.cgi?id=697092