scripts: improve git-update.sh status message
By default when the script is about to exit (normally or due to an error),
it checks whether $ERROR_LOG file exists.  If the log file exists, the
script prints a "Failures: " message prefix and dumps the log file to the
output.

Apparently the log file is always created and if the update/build is
successful, the script finishes with a bit misleading "Failures: " message.

An improvement provided with this change lets the log file to be created as
needed, i.e. if there's an error message to be printed.  If the file
doesn't exists, the script prints a "Update done" message which clearly
indicates success.

https://bugzilla.gnome.org/show_bug.cgi?id=701177