avviddec: reset coded_width/_height before trying to open codec
If coded_width/_height is supplied, the codec might use it as the
width/height and if it is wrong can lead to segfaults or video
corruption.

This is specially harmful on renegotiation scenarios where the
resolution changed. There seems to be no specific function for reseting
the AV Context in libav, so just set it directly.

https://bugzilla.gnome.org/show_bug.cgi?id=702003